// lib/types.ts
export interface Novedad {
  id: string;
  titulo: string;
  fecha: string;
  descripcion: string;
  icono: string;
  reacciones: {
    like: number;
    love: number;
    seen: number;
  };
  rolesDestinatarios: string[]; // ej: ['rrhh', 'admin', 'tecnico']
  pin?: boolean; // para fijar arriba
}